Crate problem001

Problem 1 (Multiples of 3 and 5)

Solution detail

Using the inclusion-exclusion principle, we can calculate the required sum as s_3 + s_5 - s_15, where:

s_3 is the sum of the multiples of 3 below 1000.

s_5 is the sum of the multiples of 5 below 1000.

s_15 is the sum of the multiples of 15 (i.e. divisible by both 3 and 5) below 1000.

Each of these quantites is the sum of an arithmetic progression and can be calculated by the usual formula.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the sum of all multiples of 3 or 5 below the given limit.

sum_of_multiples
Calculate the sum of the multiples of n below the given limit, using the formula for the sum of an arithmetic progression.